improve asc codec. improve dijkstra to use fibonacci heap. filter realarray.getMinMax to avoid inf.
