+++ src/cpp/plantgl/tool/util_array.h
@@ -406,12 +406,12 @@ public:
 
   /** Returns an iterator first at the minimum value, second at the
       maximum value of \e self. */
- inline std::pair<const_iterator,const_iterator> getMinAndMax( ) const {
+ inline std::pair<const_iterator,const_iterator> getMinAndMax(bool filterinf = false) const {
     const_iterator _min = this->__A.begin();
     const_iterator _max = this->__A.begin();
     for (const_iterator _i = this->__A.begin() + 1; _i != this->__A.end(); _i++)
-      if (*_i < *_min) _min = _i;
-      else if (*_i > *_max) _max = _i;
+      if ((*_i < *_min) && (!filterinf || isfinite(*_i))) _min = _i;
+      else if ((*_i > *_max) && (!filterinf || isfinite(*_i))) _max = _i;
     return std::pair<const_iterator,const_iterator>(_min,_max);
   }
